+++ CHANGELOG.md
@@ -5,7 +5,8 @@ All notable changes to the Keyboard Macro Bata extension will be documented in t
 ### [Unreleased]
 - Fix
   - Failure to detect typed characters under some conditions.
-  - Wrong location of multi-cursor after performType under some conditions.
+  - Wrong location of multi-cursor after playback of typing under some conditions.
+  - Incorrect prediction of multi-cursor location on typing detection under some conditions.
 
 ### [0.1.0] - 2021-11-17
 - Initial release